A million and one things. Added a bunch of classes and some placeholders to draw out the structure of the app. -Added CMSconnector class (basically the "API" class) -Added Exercise class that will define a set of questions -Added ExerciseActivity to display an exercise (set of questions) -Added Question class to allow creation of Question objects for the Exercise object -Added SubtopicsActivity that will be displayed when a user chooses a topic. There are buttons there to take you to the ExerciseActivity. -Added UserProgress class that will be used to keep track of the user's progress. This class can also save a file on the device! Loading functionality not yet implemented.
Added android:tag="" to all the topic buttons under Practice section. This is for the sake of some back-end logic for text-less buttons.

Added Tharshan's code for the multiple choice questions. This was refactored in a number of ways:
-improved efficiency by removing duplicate assignments across different question types
-all GUI elements are created programmatically
-ensured the intent and using the intent is consistent with the rest of the application (most notably, String[] is now ArrayList<String>)

Added a lot of placeholder methods in ExercisesActivity along with a switch statement for all the question types. These are identified as Strings: mc, tf, dnd, t. These correspond to Multiple Choice, True or False, Drag and Drop, Typing question types respectively.

There were also some minor changes here and there, but most important ones are covered above I hope.