Fixed results dialog showing incorrect ratings. Fixed a bug where upon restarting the exercise right after completing it, without returning to the topic selection page, the exercise would not be correctly identified from saved user progress.
Most if not all of the back end code I've been involved with is now nicely commented. It should help to understand the "how" of the functionality.

ExercisesActivity:
Traded 2 methods and a lot of unnecessary loops for 1 member variable. GG WP.
Removed a member variable. We've a lot of these here.
Refactored the whole thing in short. I saved many many lines of code. The class is shorter than before, even though a lot of comments have been added.

Refactoring the ExercisesActivity also involved refactoring Exercise class, which now makes much more sense than it did shortly before the refactor. There was a lot of misplaced code and redundant searches for things.